fix: move signer to mod
