+++ src/signers/mod.rs
@@ -1,3 +1,6 @@
+use crate::error::BundlrError;
+use bytes::Bytes;
+
 #[cfg(feature = "arweave")]
 pub mod arweave;
 #[cfg(feature = "cosmos")]
@@ -7,4 +10,12 @@ pub mod ed25519;
 #[cfg(any(feature = "ethereum", feature = "erc20"))]
 pub mod secp256k1;
 
-pub mod signer;
+pub trait ToPem {}
+
+pub trait Signer {
+    fn sign(&self, message: Bytes) -> Result<Bytes, BundlrError>;
+    fn sig_type(&self) -> u16;
+    fn get_sig_length(&self) -> u16;
+    fn get_pub_length(&self) -> u16;
+    fn pub_key(&self) -> Bytes;
+}
