+++ src/verify/mod.rs
@@ -1,3 +1,13 @@
+use bytes::Bytes;
+
+use crate::error::BundlrError;
+
 pub mod file;
-// pub mod stream;
 pub mod types;
+
+pub trait Verifier
+where
+    Self: Sized,
+{
+    fn verify(pk: Bytes, message: Bytes, signature: Bytes) -> Result<bool, BundlrError>;
+}
